feat: AMD64 support

+++ src/mm/heap.c
@@ -1,8 +1,8 @@
 #include <stddef.h>
 #include <stdint.h>
 
-extern unsigned int kernel_end;
-unsigned int heap_ptr = 0;
+extern uintptr_t kernel_end;
+uintptr_t heap_ptr = 0;
 
 void heap_init(void) {
     heap_ptr = (uintptr_t)&kernel_end;
